% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrate.R
\name{narrate}
\alias{narrate}
\title{Generate Documentation from data.txt}
\usage{
narrate(datatxt, format = "markdown", output_dir = "content/dataset")
}
\arguments{
\item{datatxt}{URL of the `data.txt` file.}

\item{format}{Output format, only `"markdown"` is currently supported.}

\item{output_dir}{Output directory of documentation articles.}
}
\description{
Generate a set of markdown files based on a `data.txt` file.
}
